requirements.txt
opencv-python
scikit-learn
numpy
joblib
tqdm
streamlit
pillow
